Fix instrumentation tests and update firebase session
There we missing mocks in instrumentation tests. This wasn't notice before because, since there was no authentication, it just used the API values.

Updated the firebase authentication token to not use listeners, as they do not listen to changes when the activity process is killed.

As it was agreed, updated it to do the token request pro-actively.